add missing coords in input data for experimental plotting
